Move project registration steps into pages
