+++ app/src/pages/RegisterProject/PickNameStep.svelte
